+++ app/store/generalStore.ts
@@ -6,16 +6,23 @@ type GeneralStore = {
 
     searchQuery: string;
     setSearchQuery: (query: string) => void;
+    searchEdit: string;
+    setSearchEdit: (query: string) => void;
     
     userLocation: { latitude: number; longitude: number } | null;
     setUserLocation: (location: { latitude: number; longitude: number } | null) => void;
+
 }
 
 export const useGeneralStore = create<GeneralStore>((set) => ({
     introCompleted: false,
     setIntroCompleted: (completed) => set({ introCompleted: completed }),
+    
     searchQuery: '',
     setSearchQuery: (query) => set({ searchQuery: query }),
+
+    searchEdit: '',
+    setSearchEdit: (query) => set({ searchEdit: query }),
     
     userLocation: null,
     setUserLocation: (location) => set({ userLocation: location }),
